+++ webview-ui/src/components/common/Demo.tsx
@@ -1,24 +1,23 @@
-
 import {
-    VSCodeBadge,
-    VSCodeButton,
-    VSCodeCheckbox,
-    VSCodeDataGrid,
-    VSCodeDataGridCell,
-    VSCodeDataGridRow,
-    VSCodeDivider,
-    VSCodeDropdown,
-    VSCodeLink,
-    VSCodeOption,
-    VSCodePanels,
-    VSCodePanelTab,
-    VSCodePanelView,
-    VSCodeProgressRing,
-    VSCodeRadio,
-    VSCodeRadioGroup,
-    VSCodeTag,
-    VSCodeTextArea,
-    VSCodeTextField,
+	VSCodeBadge,
+	VSCodeButton,
+	VSCodeCheckbox,
+	VSCodeDataGrid,
+	VSCodeDataGridCell,
+	VSCodeDataGridRow,
+	VSCodeDivider,
+	VSCodeDropdown,
+	VSCodeLink,
+	VSCodeOption,
+	VSCodePanels,
+	VSCodePanelTab,
+	VSCodePanelView,
+	VSCodeProgressRing,
+	VSCodeRadio,
+	VSCodeRadioGroup,
+	VSCodeTag,
+	VSCodeTextArea,
+	VSCodeTextField,
 } from "@vscode/webview-ui-toolkit/react"
 import { FormattedMessage, useIntl } from 'react-intl';
 
@@ -83,7 +82,7 @@ function Demo() {
 					))}
 				</VSCodeDataGrid>
 
-                <VSCodeTextField>
+				<VSCodeTextField>
 					<section slot="end" style={{ display: "flex", alignItems: "center" }}>
 						<VSCodeButton appearance="icon" aria-label="Match Case">
 							<span className="codicon codicon-case-sensitive"></span>